Add files via upload
In this project redeemed Heap Memory Manager. He will manage everything to do with the requests of dynamic memory allocation, and will serve as a substitute for operating system approaches. It will affect the entire dynamic allocations, for all types of possible allocation on the heap.
The project consists of several parts:
1 .ctibt infrastructure layer - which will take care of the heap memory management itself.
2. The exercise of writing a new operator operator, operator new [] operator delete [] operator delete